requirements.txt
appdirs==1.4.3
dj-database-url==0.4.2
Django==1.11
djangorestframework==3.6.2
ecdsa==0.13
future==0.16.0
gunicorn==19.7.1
packaging==16.8
psycopg2==2.7.1
pycrypto==2.6.1
PyJWT==1.5.0
pyparsing==2.2.0
python-jose==1.3.2
pytz==2017.2
requests==2.13.0
six==1.10.0
whitenoise==3.3.0
